adlpq: ADCS ADL-Prevention Questionnaire for Participant

ATRI clinical analysis dataset.

Usage

data(adlpq)

Format

A data frame with 15334 rows and 45 variables.

Details

SUBSTUDY: NA

BID: NA

VISCODE: NA

DONE: Was assessment/procedure done?

NDREASON: If No, reason not done:

ASMETHOD: Method of Collection:

ASCBOOK: 1. In the past 3 months, how well did you balance a credit card statement, pay bills, or use an ATM to make a withdrawal or deposit?

ASTRAVL: 2. In the past 3 months, did you drive a car without getting lost, or did you travel to wherever you needed by using public or other transport?

ASAPPLI: 3. In the past 3 months, did you use an appliance (for example, a microwave, dishwasher, vacuum cleaner, or TV remote), or did you carry out household repairs?

ASLAUN: 4. In the past 3 months, did you do the laundry?

ASPAY: 5. In the past 3 months, did you select and pay for items when shopping?

ASSNACK: 6. In the past 3 months, did you prepare meals or snacks?

ASTIMES: 7. In the past 3 months, how well did you remember important dates and times, such as appointments or meetings?

ASFIND: 8. In the past 3 months, did you usually find your personal belongings at home - for example keys, glasses, wallet or books?

ASWDOWN: 9. In the past 3 months, how well were you able to write things down so that other people understood them (for example, letters or detailed notes or lists)?

ASMOVIE: 10. In the past 3 months, how well did you follow TV programs or movies and remember the details of the story?

ASCEVNT: 11. In the past 3 months, how well did you talk about and remember current events that you heard or read about?

ASPHONE: 12. In the past 3 months, how well did you make telephone calls, including look up numbers or call directory assistance if necessary?

ASMEDS: 13. In the past 3 months, did you take medications regularly - including prescription or over the counter medications?

ASPLAN: 14. In the past 3 months, did you plan and organize complex activities for you or for groups of people, for example travel, running errands, participating in group activities (such as religious services), or carrying out hobbies or pastimes?

ASCOMPL: 15. In the past 3 months, did you complete complex activities such as hobbies or pastimes, for example playing card games, reading books, gardening?

ASCELL: 16. Did you ever use a cellphone or smartphone?

ASCELLUSE: 16a. If yes, during the past 3 months, did you use a cellphone or smartphone?

ASCALL: 16b.i. Make a call:

ASTEXT: 16b.ii. Send a text message:

ASINTER: 16b.iii. Access the internet (e.g., websites, Facebook, etc):

ASAPP: 16b.iv. Use an App (including playing a game):

ASCELLOFTN: 16c. Did you use a cellphone or smartphone less often than usual?

ASCELITDIF: 16d. Did you use a cellphone or smartphone:

ASCOMP: 17. Did you ever use a computer, tablet, or other device with internet access?

ASACCESS: 17a. If yes, during the past 3 months, did you use a computer, tablet, or other device with internet access?

ASWEB: 17b.i. Access websites, Facebook, or email:

ASINFO: 17b.ii. Enter information on the device, e.g, made bookmarks, wrote and sent emails, posted on Facebook, etc:

ASPURCH: 17b.iii. Make purchases or reservations:

ASPASS: 17b.iv. Use passwords and remember or find them on your own:

ASCOMPOFTN: 17c. Did you use a computer, tablet, or other device with internet access less often than usual?

ASCLITDIF: 17d. Did you use a computer, tablet, or other device with internet access:

ASEREAD: 18. Did you ever use an e-reader, e.g. Nook, Kindle or a tablet?

ASEREAD4: 18a. If yes, during the past 3 months, did you use an e-reader, e.g. Nook, Kindle or a tablet?

ASEBOOK: 18b.i. Read a book or article:

ASEDLOAD: 18b.ii. Download reading material:

ASEADJUST: 18b.iii. Make an adjustment to appearance e.g., inserted a bookmark or adjusted settings:

ASEOFTEN: 18c. Did you use an e-reader less often than usual?

ASELITDIF: 18d. Did you use an e-reader:

ASSCORE: Total Score

Examples

if (FALSE) { # \dontrun{
browseVignettes('A4LEARN')
} # }