adlpqsp: ADCS ADL-Prevention Questionnaire for Study Partner

ATRI clinical analysis dataset.

Usage

data(adlpqsp)

Format

A data frame with 15523 rows and 88 variables.

Details

SUBSTUDY: NA

BID: NA

VISCODE: NA

DONE: Was assessment/procedure done?

NDREASON: If No, reason not done:

AIMETHOD: Method of Collection:

BPID: Study Partner ID:

AICBOOK: 1. In the past 3 months, how well did {P} balance a credit card statement, pay bills, or use an ATM to make a withdrawal or deposit?

AISUB1A: If with some difficulty:

AISUB1B: NA

AISUB1C: NA

AITRAVL: 2. In the past 3 months, did {P} drive a car without getting lost, or travel to wherever he/she needed to go by using public or other transport?

AISUB2A: If with some difficulty:

AISUB2B: NA

AISUB2C: NA

AIAPPLI: 3. In the past 3 months, how well did {P} use an appliance (for example, a microwave, dishwasher, vacuum cleaner, or TV remote), or carry out household repairs?

AISUB3A: If with some difficulty:

AISUB3B: NA

AISUB3C: NA

AILAUN: 4. In the past 3 months, did {P} clean a load of laundry?

AISUB4A: If with some difficulty:

AISUB4B: NA

AISUB4C: NA

AIPAY: 5. In the past 3 months, did {P} select and pay for items when shopping?

AISUB5A: If with some difficulty:

AISUB5B: NA

AISUB5C: NA

AISNACK: 6. In the past 3 months, did {P} prepare a meal or snack?

AISUB6A: If with some difficulty:

AISUB6B: NA

AISUB6C: NA

AITIMES: 7. In the past 3 months, how well did {P} remember important dates and times, such as appointments or meetings?

AISUB7A: If with some difficulty:

AISUB7B: NA

AISUB7C: NA

AIFIND: 8. In the past 3 months, did {P} usually manage to find personal belongings at home - for example keys, glasses, a wallet or books?

AISUB8A: If with some difficulty:

AISUB8B: NA

AISUB8C: NA

AIWDOWN: 9. In the past 3 months, how well did {P} write things down, for example letters or detailed notes or lists, so that other people understood them?

AISUB9A: If with some difficulty:

AISUB9B: NA

AISUB9C: NA

AIMOVIE: 10. In the past 3 months, how well did {P} follow TV programs or movies and remember the details of the story?

AISUB10A: If with some difficulty:

AISUB10B: NA

AICEVNT: 11. In the past 3 months, how well did {P} talk about and remember current events that he/she heard or read about?

AISUB11A: If with some difficulty:

AISUB11B: NA

AIPHONE: 12. In the past 3 months, how well did {P} make telephone calls, including looking up numbers or calling directory assistance if necessary?

AISUB12A: If with some difficulty:

AISUB12B: NA

AISUB12C: NA

AIMEDS: 13. In the past 3 months, did {P} take medications regularly - including prescription or over the counter medications?

AISUB13A: If with some difficulty:

AISUB13B: NA

AIPLAN: 14. In the past 3 months, did {P} plan and organize complex activities for him/herself or for groups of people, for example travel, running errands, participating in group activities (such as religious services), or carrying out hobbies or pastimes?

AISUB14A: If with some difficulty:

AISUB14B: NA

AISUB14C: NA

AICOMPL: 15. In the past 3 months, did {P} complete complex activities such as hobbies or pastimes, for example playing card games, reading books, gardening?

AISUB15A: If with some difficulty:

AISUB15B: NA

AISUB15C: NA

AICELL: 16. Did {P} ever use a cellphone or smartphone?

AICELLUSE: 16a. If yes, during the past 3 months, did {P} use a cellphone or smartphone?

AICALL: 16b.i. Make a call:

AITEXT: 16b.ii. Send a text message:

AIINTER: 16b.iii. Access the internet (e.g., websites, Facebook, etc):

AIAPP: 16b.iv. Use an App (including playing a game):

AICELLOFTN: 16c. Did {P} use a cellphone or smartphone less often than usual?

AICELITDIF: 16d. Did {P} use a cellphone or smartphone:

AICOMP: 17. Did {P} ever use a computer, tablet, or other device with internet access?

AIACCESS: 17a. If yes, during the past 3 months, did {P} use a computer, tablet, or other device with internet access?

AIWEB: 17b.i. Access websites, Facebook, or email:

AIINFO: 17b.ii. Enter information on the device, e.g, made bookmarks, wrote and sent emails, posted on Facebook, etc:

AIPURCH: 17b.iii. Make purchases or reservations:

AIPASS: 17b.iv. Use passwords and remember or find them on his/her own:

AICOMPOFTN: 17c. Did {P} use a computer, tablet, or other device with internet access less often than usual?

AICLITDIF: 17d. Did {P} use a computer, tablet, or other device with internet access:

AIEREAD: 18. Did {P} ever use an e-reader, e.g. Nook, Kindle or a tablet?

AIEREAD4: 18a. If yes, during the past 3 months, did {P} use an e-reader, e.g. Nook, Kindle or a tablet?

AIEBOOK: 18b.i. Read a book or article:

AIEDLOAD: 18b.ii. Download reading material:

AIEADJUST: 18b.iii. Make an adjustment to appearance e.g., inserted a bookmark or adjusted settings:

AIEOFTEN: 18c. Did {P} use an e-reader less often than usual?

AIELITDIF: 18d. Did {P} use an e-reader:

AISCORE: Total Score

Examples

if (FALSE) { # \dontrun{
browseVignettes('A4LEARN')
} # }