cfi: Cognitive Function Index - Participant

ATRI clinical analysis dataset.

Usage

data(cfi)

Format

A data frame with 15354 rows and 22 variables.

Details

SUBSTUDY: NA

BID: NA

VISCODE: NA

DONE: Was assessment/procedure done?

NDREASON: If No, reason not done:

PMETHOD: Method of Collection:

MEMORY: 1. Compared to one year ago, do you feel that your memory has declined substantially?

REPEAT: 2. Do others tell you that you tend to repeat questions over and over?

MISPLA: 3. Have you been misplacing things more often?

WRITTN: 4. Compared to one year ago, do you find that you are relying more on written reminders (e.g., shopping lists, calendars)?

HELP: 5. Do you need more help from others to remember appointments, family occasions or holidays?

RECALL: 6. Do you have more trouble recalling names, finding the right word, or completing sentences?

DRIVE: 7. Do you have more trouble driving (e.g., do you drive more slowly, have more trouble at night, tend to get lost, have accidents)?

MONEY: 8. Compared to one year ago, do you have more difficulty managing money (e.g., paying bills, calculating change, completing tax forms)?

SOCIAL: 9. Are you less involved in social activities?

WORK: 10. Has your work performance (paid or volunteer) declined significantly compared to one year ago?

FOLLOW: 11. Do you have more trouble following the news or the plots of books, movies, or TV shows?

ACTIV: 12. Are there any activities (e.g., hobbies, such as card games, crafts) that are substantially more difficult for you now compared to one year ago?

LOST: 13. Are you more likely to become disoriented, or get lost, for example when traveling to another city?

APPLIA: 14. Compared to one year ago, do you have more difficulty using household appliances (such as washing machine or microwave) or electronic devices (such as computer or cell phone)?

PTCONCERN: 15. In the past year, have you seen a doctor about memory concerns?

CFIPTTOTAL: Total Score

Examples

if (FALSE) { # \dontrun{
browseVignettes('A4LEARN')
} # }