cfisp: Cognitive Function Index - Study Partner

ATRI clinical analysis dataset.

Usage

data(cfisp)

Format

A data frame with 15539 rows and 23 variables.

Details

SUBSTUDY: NA

BID: NA

VISCODE: NA

DONE: Was assessment/procedure done?

NDREASON: If No, reason not done:

BPID: Study Partner ID:

SPMETHOD: Method of Collection:

SPMEMORY: 1. Do you feel that the participant has had a significant decline in memory compared to one year ago?

SPREPEAT: 2. Does the participant tend to ask the same question over and over?

SPMISPLA: 3. Has the participant been misplacing things more often?

SPWRITTN: 4. Compared to one year ago, does it seem to you that the participant is relying more on written reminders (e.g., shopping lists, calendars)?

SPHELP: 5. Does the participant need more help from others to remember appointments, family occasions or holidays?

SPRECALL: 6. Does the participant have more trouble recalling names, finding the right word, or completing sentences?

SPDRIVE: 7. Is the participant having more trouble driving (e.g., driving too slowly or too fast, having more trouble driving at night, getting lost, having accidents)?

SPMONEY: 8. Compared to one year ago, is the participant having more difficulty managing money (e.g., paying bills, calculating change, completing tax forms)?

SPSOCIAL: 9. Is the participant less interested in social activities?

SPWORK: 10. Do you believe, based on your own observations or comments from the participant's co-workers, that the participant's work performance (paid or volunteer) has declined significantly, compared to one year ago?

SPFOLLOW: 11. Does the participant have more trouble following the news or the plots of books, movies, or TV shows?

SPACTIV: 12. Are there any activities (e.g., hobbies such as card games, crafts) that are substantially more difficult for the participant now compared to one year ago?

SPLOST: 13. Is the participant more likely to become disoriented, or get lost, for example when traveling to another city?

SPAPPLIA: 14. Compared to one year ago, does the participant have more difficulty using household appliances (such as washing machine or microwave) or electronic devices (such as computer or cell phone)?

SPCONCERN: 15. In the past year, has the participant seen a doctor about memory concerns?

CFSPTTOTAL: Total Score

Examples

if (FALSE) { # \dontrun{
browseVignettes('A4LEARN')
} # }