mmse: Mini Mental State Exam

ATRI clinical analysis dataset.

Usage

data(mmse)

Format

A data frame with 26765 rows and 41 variables.

Details

SUBSTUDY: NA

BID: NA

VISCODE: NA

DONE: Was assessment/procedure done?

NDREASON: If No, reason not done:

MMSELOC: Where was the assessment done?

WORDLIST: Which word list was used?

MMDATE: 1. What is today's date?

MMYEAR: 2. What is the year?

MMMONTH: 3. What is the month?

MMDAY: 4. What day of the week is today?

MMSEASON: 5. What season is it?

MMHOSPIT: 6. What is the name of this hospital (clinic, place)?

MMFLOOR: 7. What floor are we on?

MMCITY: 8. What town or city are we in?

MMAREA: 9. What county (district, borough, area) are we in?

MMSTATE: 10. What state are we in?

MMBALL: 11. Word 1

MMFLAG: 12. Word 2

MMTREE: 13. Word 3

MMTRIALS: 13a. Enter number of trials

MMDLTR: 14. Indicate first letter

MMLLTR: 15. Indicate second letter

MMRLTR: 16. Indicate third letter

MMOLTR: 17. Indicate fourth letter

MMWLTR: 18. Indicate fifth letter

MMWORLD: Attention sub-score (WORLD Backwards - CERAD scoring method)

MMBALLDL: 19. Word 1

MMFLAGDL: 20. Word 2

MMTREEDL: 21. Word 3

MMWATCH: 22. Show the participant a wrist watch and ask "What is this?"

MMPENCIL: 23. Repeat for pencil.

MMREPEAT: 24. Say, "Repeat after me: no ifs, ands, or buts."

MMHAND: 25. Takes paper in right hand.

MMFOLD: 26. Folds paper in half.

MMONFLR: 27. Puts paper on floor.

MMREAD: 28. Present the piece of paper which reads, "CLOSE YOUR EYES," and say: "Read this and do what it says."

MMWRITE: 29. Give the participant a blank piece of paper and say: "Write a sentence."

MMDRAW: <big header>Construction</big header><p>30. Present the participant with the Construction Stimulus page. Say, "Copy this design."</p>

MMSCORE: MMSE TOTAL SCORE

MMSE: MMSE (PACC Audio Verification)

Examples

if (FALSE) { # \dontrun{
browseVignettes('A4LEARN')
} # }